+++ tests/common/mod.rs
@@ -12,7 +12,7 @@ use dnsimple_rust::dnsimple::{Client, new_client};
 /// `fixture`: the path to the fixture inside the `api` directory
 /// `path`: the path in the server (i.e. `/whoami`)
 ///
-pub fn setup_mock_for(path: &str, fixture: &str) -> (Client, Mock) {
+pub fn setup_mock_for(path: &str, fixture: &str, method: &str) -> (Client, Mock) {
     let path = format!("/v2{}", path);
     let fixture = format!("./tests/fixtures/v2/api/{}.http", fixture);
     println!("We are trying to read this file: {}", fixture);
@@ -21,12 +21,15 @@ pub fn setup_mock_for(path: &str, fixture: &str) -> (Client, Mock) {
         .expect("Something went wrong: Couldn't read the file");
     let tokens = content.split("\r\n\r\n");
     let vec = tokens.collect::<Vec<&str>>();
+    let headers = vec.first().unwrap();
+    let status = &headers[9..12];
     let body = vec.last();
 
-    let mock = mock("GET", path.as_str())
+    let mock = mock(method, path.as_str())
         .with_header("X-RateLimit-Limit", "2")
         .with_header("X-RateLimit-Remaining", "2")
         .with_header("X-RateLimit-Reset", "never")
+        .with_status(status.parse().unwrap())
         .with_body(body.unwrap()).create();
 
     let mut client = new_client(true, String::from("some-token"));