+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 ## main
+
+- ADDS: OAuth Service and Error handling
+- ADDS: Accounts Service
+- ADDS: Identity Service
+- ADDS: General Client Structure